requirements.txt
django
djangorestframework
pytesseract
pillow
gunicorn
numpy
opencv-python-headless
